requirements.txt
langchain_core==0.2.5
langchain_groq==0.1.5
python-dotenv==1.0.1
streamlit==1.35.0

mysql-connector-python==8.3.0
SQLAlchemy==2.0.27
pymysql==1.1.0
st_copy_to_clipboard==0.1.6
requests==2.31.0
#streamlit==1.32.0
st-annotated-text==4.0.1

pypdf2==3.0.1
transformers==4.41.2